bugg igen

+++ bombDlg.h
@@ -212,8 +212,8 @@ public:
 	afx_msg void OnBnClickedButton4();
 	afx_msg void OnBnClickedRadio17();
 	BOOL m_en_krona_ospelad;
-	double m_maxnomodds;
-	double m_minodds;
+	UINT m_maxnomodds;
+	UINT m_minodds;
 };
 
 //{{AFX_INSERT_LOCATION}}
